+++ examples/hybrid.py
@@ -1,0 +1,4 @@
+import cspark.wasm as Hybrid
+
+response = Hybrid.Client.health_check()
+print(response.data)